+++ crates/ruma-common/src/api/error.rs
@@ -41,6 +41,20 @@ impl Error {
     pub fn error_kind(&self) -> Option<&ErrorKind> {
         as_variant!(&self.body, ErrorBody::Standard(StandardErrorBody { kind, .. }) => kind)
     }
+
+    /// Whether this error matches the expected format for an endpoint that is not implemented by
+    /// the homeserver.
+    ///
+    /// Return `true` if this contains an [`ErrorKind::Unrecognized`] with a
+    /// [`http::StatusCode::NOT_FOUND`].
+    ///
+    /// [unsupported endpoint]:
+    pub fn is_endpoint_not_implemented(&self) -> bool {
+        self.status_code == http::StatusCode::NOT_FOUND
+            && self
+                .error_kind()
+                .is_some_and(|error_kind| matches!(error_kind, ErrorKind::Unrecognized))
+    }
 }
 
 impl fmt::Display for Error {